coll/han: fix compiler warnings
Fix an unused variable, and explictly cast uint64_t to size_t

Note that the latter won't be needed after embiggening collectives for large count support.

+++ ompi/mca/coll/han/coll_han_utils.c
@@ -25,12 +25,12 @@
  * @param[in]   size            Number of numerators
  * @returns     The GCD, where 1 <= GCD
  */
-size_t coll_han_utils_gcd(const size_t *numerators, const size_t size)
+size_t coll_han_utils_gcd(const uint64_t *numerators, const size_t size)
 {
-    size_t denominator = numerators[0], numerator, tmp;
+    size_t denominator = (size_t) numerators[0], numerator, tmp;
 
     for (size_t i = 1; i < size; ++i) {
-        numerator = numerators[i];
+        numerator = (size_t) numerators[i];
 
         if (0 == denominator) {
             denominator = numerator;
